+++ src/app/src/App.js
@@ -71,16 +71,7 @@ function PrivateRoutes() {
                 <Route path='/submissions/*' element={<Submissions />} />
                 <Route
                     path='*'
-                    element={
-                        <Navigate
-                            to={
-                                hasWelcomePageAccess
-                                    ? '/welcome'
-                                    : '/submissions'
-                            }
-                            replace
-                        />
-                    }
+                    element={<Navigate to={'/submissions'} replace />}
                 />
             </Routes>
         </>