+++ src/app/src/components/Submissions/List.js
@@ -1,3 +1,4 @@
+import { useEffect } from 'react';
 import {
     HStack,
     Box,
@@ -27,7 +28,7 @@ import {
     LocationMarkerIcon,
 } from '@heroicons/react/solid';
 import { useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 
 import { heroToChakraIcon } from '../../utils';
 import { useGetBoundariesQuery } from '../../api/boundaries';
@@ -37,13 +38,35 @@ import { ROLES } from '../../constants';
 
 export default function SubmissionsList() {
     const navigate = useNavigate();
+    const location = useLocation();
+    const user = useSelector(state => state.auth.user);
+    const utilityId = useSelector(state => state.auth.utility?.id);
+
+    const {
+        isFetching,
+        data: boundaries,
+        error,
+    } = useGetBoundariesQuery({
+        utilities: user.role === ROLES.CONTRIBUTOR ? utilityId : undefined,
+    });
+
+    // Navigate to /welcome if coming from /login and no boundaries already
+    useEffect(() => {
+        if (boundaries?.length === 0 && location.state?.pathname === '/login') {
+            navigate('/welcome');
+        }
+    }, [location.state, boundaries, navigate]);
 
     return (
         <Box paddingLeft={8} paddingRight={8} mt={6}>
             <Flex>
                 <Heading size='lg'>Submissions</Heading>
                 <Spacer />
-                <Button mr={4} onClick={() => navigate('/welcome')}>
+                <Button
+                    mr={4}
+                    disabled={boundaries?.length > 0}
+                    onClick={() => navigate('/welcome')}
+                >
                     Add map
                 </Button>
             </Flex>
@@ -59,7 +82,11 @@ export default function SubmissionsList() {
                             <Table variant='submissions'>
                                 <TableHeader />
                                 <Tbody>
-                                    <TableRows />
+                                    <TableRows
+                                        isFetching={isFetching}
+                                        error={error}
+                                        boundaries={boundaries}
+                                    />
                                 </Tbody>
                             </Table>
                         </TableContainer>
@@ -99,18 +126,7 @@ function TableHeader() {
     );
 }
 
-function TableRows() {
-    const user = useSelector(state => state.auth.user);
-    const utilityId = useSelector(state => state.auth.utility?.id);
-
-    const {
-        isFetching,
-        data: boundaries,
-        error,
-    } = useGetBoundariesQuery({
-        utilities: user.role === ROLES.CONTRIBUTOR ? utilityId : undefined,
-    });
-
+function TableRows({ isFetching, error, boundaries }) {
     if (isFetching) {
         return <LoadingRow />;
     }