+++ src/django/api/serializers/reference_image.py
@@ -17,3 +17,15 @@ class ReferenceImageSerializer(ModelSerializer):
             # DRF already validates negative values for PositiveIntegerField
             raise ValidationError("Opacity takes values from 0-100.")
         return value
+
+
+class ReferenceImageUpdateSerializer(ReferenceImageSerializer):
+    class Meta:
+        model = ReferenceImage
+        exclude = ("boundary",)
+        read_only_fields = [
+            "uploaded_at",
+            "uploaded_by",
+            "file",
+            "filename",
+        ]