+++ src/django/api/tests/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa: F401
+from .base import BoundarySyncAPITestCase
+from .boundary import ShapeUploadTests
+from .reference_image import ReferenceImageViewTests