+++ src/django/iow/urls.py
@@ -19,9 +19,13 @@ from django.conf.urls.static import static
 
 from . import settings
 
-urlpatterns = [
-    path('admin/', admin.site.urls),
-    path('api/', include('api.urls')),
-    path('health-check/', include('watchman.urls')),
-    path('confirm_password_reset/', include('django.contrib.auth.urls')),
-] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
+urlpatterns = (
+    [
+        path("admin/", admin.site.urls),
+        path("api/", include("api.urls")),
+        path("health-check/", include("watchman.urls")),
+        path("confirm_password_reset/", include("django.contrib.auth.urls")),
+    ]
+    + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
+    + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+)